backend/requirements.txt
torch==2.1.0
torchvision==0.16.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
yfinance==0.2.18
newsapi-python==0.2.6
python-dotenv==1.0.0
ta==0.10.2
plotly==5.15.0